Converted equipment detail page to use async
The problem with the OEE Dial appears to be related to the data availability on page load, this makes it better but there are still strange issues occuring.
What I thought was originally the data not being present on page load was actually vue-router stalling and no route being selected.